@@ -4,16 +4,17 @@
   "main": "index.js",
   "license": "MIT",
   "dependencies": {
-    "next": "^6.1.1",
+    "next": "^8.0.0-canary.2",
     "prop-types": "^15.6.2",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0"
+    "react": "16.7.0-alpha.0 - next",
+    "react-dom": "16.7.0-alpha.0 - next"
   },
   "scripts": {
     "dev": "next",
-    "build": "next build && next export",
-    "start": "serve -p 3000 out",
-    "deploy": "now && now alias"
+    "build": "next build",
+    "start": "next",
+    "deploy": "now && now alias",
+    "test": "echo 'No tests' && exit 0"
   },
   "devDependencies": {
     "eslint": "^4.19.1",
@@ -21,7 +22,6 @@
     "eslint-config-prettier": "^2.9.0",
     "eslint-plugin-import": "^2.13.0",
     "eslint-plugin-jsx-a11y": "^6.1.1",
-    "eslint-plugin-react": "^7.10.0",
-    "serve": "^10.0.2"
+    "eslint-plugin-react": "^7.10.0"
   }
 }
